pyproject.toml
[project]
name = "python-config-loader"
description = "A module to load and merge configuration files from multiple formats."

# If a project, then you can specify version here
# version = "0.0.0"
# If a module, then fetch version dynamically (see [tool.setuptools.dynamic] below)
dynamic = ["version"]

readme = "README.md"
requires-python = ">=3.9"
license = {file = "LICENSE"}
authors = [
  {email = "matthew@davidson.engineering"},
  {name = "Matthew Davidson"}
]

classifiers = [
    "Development Status :: 1 - Planning",
    "Operating System :: Microsoft :: Windows",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
   "pyyaml >= 6.0.0",
]

# If a module, then use fetch version from src/module_name/__init__.py
[tool.setuptools.dynamic]
version = {attr = "config_loader.__version__"}

[project.optional-dependencies]
test = [
  "pytest >= 7.1.1",
]
